mission_process/requirements.txt
sqlalchemy
kafka-python
psycopg2-binary
